// web/lib/api/generated/model/platformTextSuggestionsListStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type PlatformTextSuggestionsListStatus = typeof PlatformTextSuggestionsListStatus[keyof typeof PlatformTextSuggestionsListStatus];


export const PlatformTextSuggestionsListStatus = {
  NEW: 'NEW',
  DONE: 'DONE',
  DISMISSED: 'DISMISSED',
  '': '',
} as const;